Rename checks for better consistency
Hopefully this will help reduce confusion.
Conversations are hard when each check has to be explained each time -
this tries to make names more obvious.
