+++ lib/checks/README.md
@@ -17,27 +17,27 @@ However, they may have overlapping output, in that several checks may detect the
 
 ## Checks
 
-### BasePathsMissingFromPublishingApi
+### RummagerContentMissingFromPublishingApi
 
-This check finds content present in Rummager for which the base_path doesn't map to a published content item in the Publishing API.
-The assumption is that anything present in Rummager should be published.
-Our proxy for 'being published' is that the `/lookup-by-base-path` endpoint in Publishing API finds a content_id.
+This check finds content present in Rummager for which the base_path doesn't map to a user-visible content item in the Publishing API.
+The assumption is that anything present in Rummager should be user-visible.
+Our proxy for 'user-visible' is that the `/lookup-by-base-path` endpoint in Publishing API finds a content_id.
 Recommended links are excluded.
 Withdrawn items are excluded, because the content_id lookup does not guarantee to find a content_id for them.
 
-### BasePathsMissingFromRummager
+### PublishingApiContentMissingFromRummager
 
-This check finds published content present in Publishing API for which no base_path is present in Rummager.
+This check finds user-visible content present in Publishing API for which no base_path is present in Rummager.
 The assumption is that everything published in Publishing API should be in Rummager.
-Our proxy for 'being published' is that the some content item of the content_id is state `published`, `unpublished`, or `superseded`.
+Our proxy for 'being published' is that the some content item of the content_id has state `published`, `unpublished`, or `superseded`.
 
-### LinkedBasePathsMissingFromPublishingApi
+### RummagerLinkTargetsMissingFromPublishingApi
 
-This check finds link targets for links present in Rummager for which the target base_path doesn't map to a published content item in the Publishing API.
-The assumption is that any link target in Rummager should be published.
-Our proxy for 'being published' is that the `/lookup-by-base-path` endpoint in Publishing API finds a content_id.
+This check finds link targets for links present in Rummager for which the target base_path doesn't map to a user-visible content item in the Publishing API.
+The assumption is that any link target in Rummager should be user-visible.
+Our proxy for 'user-visible' is that the `/lookup-by-base-path` endpoint in Publishing API finds a content_id.
 
-### RummagerLinksNotIndexedInRummager
+### RummagerLinkTargetsMissingFromRummager
 
 This check finds link targets for links present in Rummager for which the target base_path is not present in Rummager.
 The assumption is that any link target in Rummager should be indexed.
@@ -52,7 +52,7 @@ The assumption is that links should always be in sync.
 This check finds links present in Publishing API which are not present in Rummager.
 The assumption is that links should always be in sync.
 
-# RummagerRedirects
+### RedirectedRummagerContent
 
 Checks for documents in rummager where the base path is now redirected according to publishing api.
 
@@ -61,7 +61,7 @@ This is equivalent to the following:
 - Lookup the base path of the rummager document to get a content id
 - Check whether a redirect exists for that content id
 
-# RummagerRedirectedLinks
+### RedirectedRummagerLinkTargets
 
 Checks for links in rummager that point to content that is redirected in the publishing api.
