+++ lib/checks/redirected_rummager_content.rb
@@ -1,5 +1,5 @@
 module Checks
-  class RummagerRedirects
+  class RedirectedRummagerContent
     def initialize(name, checker_db, reporter)
       @name = name
       @checker_db = checker_db
